Fix some issues that prevents CNF Cert Suite and related libvirt tests from running correctly
This change is gathering several small changes that are really needed for new CNF Cert Suite release, also to meet some partners requirements and for example-cnf.
All the possible changes are included in this single CR instead of creating several CRs because CNF Cert Suite execution on d-o-a-a is currently broken, so that this CR will cover the fixes for the errors we are having now plus the improvements to minimize the failed test suites we are having now.
List of changes:
-Multiple autodiscovery labels is now supported, using lists instead of bare variables for tnf_targetpodlabels variable.
-As this feature already selects the pods under test by the CNF Cert Suite, tnf_cnfs_regexp variable is no longer needed, so it is deleted.
-Debug daemonSet, currently created by the tnf container, is now supported. The change includes the logic to copy the image in disconnected environments, to properly define REDHAT_RHEL_REGISTRY env var and to remove it after the execution if required
-To remove debug daemonSet, the same variable used for partner deployment is used, renamed to tnf_postrun_delete_resources for a better understanding
-Including tnf_log_level variable to tune the log level used for executing CNF Cert Suite
-During the tests, it was found that the latest version of tnf container was not being used and the cache was being used, so it will be always downloaded to have the latest changes
-Also including a change in cnf-cert post-run task to use the correct oc path because it was causing failures in libvirt checks
-Changing tnf_test_example to meet with the modifications commented above.
-Finally, some typos corrected

Change-Id: Idd10029973c68d986a4fb50d97e44ece92301c27
Test-App-Hints: tnf_test_example
Test-Args-Hints: -e dci_openshift_app_ns=test-check -e do_cnf_cert=true -e tnf_suites=diagnostic