Mirror image used in basic_example sample if disconnected env
Right now, we are supposing that the image used in basic_example sample is present in the registry if we are using it, but it may not be the case.
This change applies the copy of the image in case of having a disconnected environment.

Moreover, changing chart-verifier README to remove references to a specific port registry

Change-Id: I09e061c7febb9780c1e06e0f18bbd35c64c68759