Inserting boot iso in a VM might be slow
Effectively, the sushy redfish emulator would copy the disk you upload
to a temp location, depending on the size of the boot iso this could
take quite a bit of time and space since the loop will retry to insert
the iso again and again, exhausting the space in /tmp.
Increasing the timeout from the default 30s to 90s seems to yield
consistent success.