Add hash caching and simplify overriding assisted installer images
Hashes are now saved to a file in the repo root. If the
image key is already in the image_hashes saved we will not try
to get a new one. If you change version you expect to get you
may want to remove the value from the file or the file completely.
There is also a flag `ignore_cached_image_hashes` which will
ignore any save values and write a new file.

It is now much simpler to override the values in the os_images and
release_images as the format is now the same as assisted installer
expects. And there is no need to override as mean values as previous
just override os_images (if required), release_images and
supported_ocp_versions.