increase time to wait for worker nodes to be ready
sometimes workers are blocked because of a pending
mc to apply and when applying the work-around to
uncordon the node, but if the pending mc required
a reboot, it takes more than 300 seconds for the
node to be in a running state. This change increases
the time to validate a node to be ready to 600 secs
to consider those cases.

Test-Hints: sno
Change-Id: Ib3478e44c190f3d11895a8f0ba5381cae6c03387