Make job to fail if CNF Cert Suite execution did not finish correctly
Currently, we are not making the job to fail if CNF Cert Suite execution is not correctly finished. This code would help to anticipate this kind of issues and not make the job fail in other stages of the agent that depends on the CNF Cert Suite, e.g. sync-cnf-cert-and-preflight role

Change-Id: Iab02cfc65d8b73c32efc4211ca4f5f870e5fb3d3