Remove tnf image in teardown and only if it is not the latest stable version
Right now, the tnf image is removed in post-run, but it may happen that the tnf execution did not work fine and then the failure process is triggered before the post-run, so that the image will never be deleted. Moving it to teardown instead, to ensure it is deleted.

Also, to avoid collisions during daily jobs, as latest stable version is tested in several daily jobs, only remove the images from versions different than the latest stable one (e.g. main branch or PRs)

Change-Id: I04a940edf36356e56300308d765de39d1d87b3fc