dci-openshift-app-agent-ctl - avoid changing config_dir and config in dev mode
In the current logic, if we run d-o-a-a-ctl in dev mode (-d), it is explicitly setting config_dir and config variables to point to the d-o-a-a change. But this is not compatible when using prefixes, because in that case, we would like to use the original CONFIG_DIR as config_dir value to retrieve the correct settings file based on the prefix.

This change allows to only modify config if we explicitly use -c. Explaining this in the docs. Else (which it will be the case of the CI using dci-check-change and test-runner), it will use the settings file from CONFIG_DIR (default to /etc/d-o-a-a), using prefixes if provided.

- Changing the name of a task in cnf-cert role to be more explicit, and also to see if the change is correctly applied

Change-Id: I5f53521aaa089f5fd72822271c9b9c7318278efb