Wait 5 minutes in case we have nodes in NotReady,SchedulingDisabled status
There have been recent cases where we are failing when detecting a node in NotReady,SchedulingDisabled status, and we can see it in the must-gather, but when checking the nodes.txt log from DCI jobs (which is captured 2-3 minutes after the must-gather), the node moves to Ready status.

Let's wait around 5 minutes to avoid false positive issues related to the NotReady,SchedulingDisabled case.

Change-Id: I4438fc87497c093d9c815b9f078952adbdb25890