Fix cert_project_id length validation from last PR
This patch is to add cert_project_id | length > 0

Main use-case to create,update and attach helm
chart project is working fine but the issue
occurred when test with two projects are already
existed, the first project skipped the update
setting task but second project is not due to
cert_project_id set to empty so cert_project_id
is defined is still met the condition so it did not
skip the task. This patch is to fix this issue.

Change-Id: I917714f8eddc59bcf453eaf6cbca94f160a43327