Change certifiedcontainerinfo format in tnf_config.yml.j2 to adapt to changes from tnf v4.3.4
tnf changed the way this certifiedcontainerinfo field was built in this PR: https://github.com/test-network-function/cnf-certification-test/pull/1313, but the examples were not changed as well.

Now we have realised the affiliated-certification-container-is-certified test is being skipped because we're not following the new format, including since tnf v4.3.4

This change considers this update starting from that version.

Change-Id: I546f82c2b24df74e652f45d166baedb3fa0566c9