Installing conserver packages from list
Currently the conserver packages are installed by calling the "package"
module and looping through the list of packages.

The package module allows providing it with a list of packages and
installing them altogether, thus saving time and log space.

This is a small change to leverage this feature from the package module
in the conserver role.

Change-Id: Ieb906522d45c67a33ddf7b53705117fb9b58595e