Adding sushy tools to the ocp-on-libvirt solution
Currently ocp-on-libvirt installs the vBMC service along with the
VMs to allow for IPMI remote management of the hosts, which in turn
allows for PXE image boot up.

In order to support ACM managed deployments, virtual media image boot
up is needed, and to manage it remotely, the target server BMCs must
suport the Redfish protocol.

Sushy tools provides support for the Redfish protocol for libvirt VMs.

This change installs sushy tools, as implemented for
assisted-installer based deployments,  when running the ocp-on-libvirt
role.