Honor warnings from ansible-lint
CI for ansible-lint PRs is ignoring the warnings from ansible-lint.
Creating another section to highlight the warnings without failing the
check.
Including a new warning var-naming[no-role-prefix] while we decide to
use this standard or keep the current one.
Through this mechanism we can remove the warning list once we want to
start truly enforcing them.