Create a custom rule module for ansible-lint
Ansible lint allows you to write custom rules and add them to the list
of things you want to check. Adding custom rules to start enforcing
naming convention guidelines in the repo.
Also reworked `hack/ansible-lint.sh` to be used in the same way both in
GHA and in your development environment. Also the GHA is ignoring hack
directory when linting/checking

Signed-off-by: Jorge Gallegos <jgallego@redhat.com>