ACM - Move the Agent Service config to acm_setup
An agent service config is a requirement for acm_sno, acm_hypershift,
and acm_ztp roles. Moving the creating to acm_setup to allow to have
the requirements for ACM base installation during the configuration of
the hub/management cluster.

Increases time for ASC as it may take really long time in connected
environments

Co-authored-by: Tony Garcia <tonyg@redhat.com>