Add backward compatibility to el8 baremetal installer
The baremetal installer (openshift-baremetal-install) drops support to
el8 starting on 4.16. Another installer (openshift-install) provides
support to it for el8. Doing a drop-in replacement when provisioner OS
is el8 and OCP release is 4.16+

For disconnected environments, when mirror the ocp release is required,
extract both installers to be ready during the install.