Create admin user in oc_setup role
The previous admin user in the oc_setup rule was created with a
cluster-admin role. Because there is an admin role, we are now creating
three users: admin, cluster_admin, nonadmin.

nonadmin remains without change, this user has no role associated, so
nothing can be done with it, other than authenticate.

cluster_admin, is a new user and is now assigned the cluster-admin role,
previoulsy this was the role assigned to the admin user.

admin, this user now has the admin role, it has less priviledges as
before when it had cluster-admin role.