Deprecate unused tasks and simplify check_resource
The check_resource used to take actions for uncordoning nodes that is
not in used in a long time. Most of the code was to check things but not
to act on them, simplifying the role to only wait for the state of
certain kind or resources.

Deprecating the use of oc for tasks that can be performed through the k8s
module.
Removing options not in use.