Modify lookup file usage to enable remote file retrieval
This commit changes the way the lookup file is handled in the
DCI OpenShift Agent. Previously, the lookup file could only access
local files, which posed a limitation for containerized deployments.

With this update, the lookup file can now retrieve content from
a remote system, specifically allowing for the cache files to
reside on the jumphost.