create_certification_project: Fix multiple project issue
Only the first project is created and correctly linked to
the corresponding product ID. However, subsequent projects
are not created because the project ID from the first entry
in the certification project job was not reset.

As a result, this initial project ID is incorrectly associated
with all product list IDs.