Multiple fixes to run testing locally
This fixes the issue of running on a detached branch, i.e. using
HEAD, which is the same used for the current change.
Uses --lint for sanity tests, as newer versions send output to stdout
instead of stderr which is parser to catch differences between main and
the branch checked.
Removes unshallow fetch as we are not dealing with bare repositories.