sno_node_prep: take cache server as an input var
Before this change, the cache_server identity was one of:

- the inventory provision_host
- the inventory registry_host

Depending on whether the cache_provisioner variable was set to true or false.

The sno_node_prep role set this variable based on the result of a check to access the Internet from the provision_host so:

- with internet access, all the operations involving downloading, extracting, processing, serving and running the installation artifacts was done from the provision host.

- with no internet access, the downloading, extracting, processing and serving tasks were delegated to the registry host, and the required artifacts were copied to the provision_host for running.

This posed a problem for environments where the provision host had access to the Internet but didn't act as the cache server.

This change removes the logic described above with the simpler:

- By default, the provision_host is used as the cache server.

- If this is not the desired behavior, the snp_cache_server and si_cache_server variables may be set from the inventory to the identity of the desired cache server.

This entails knowing in advance whether the provision host is suitable to be the cache server (does it have Internet access and enough disk space?).